backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Configuration management
pydantic==2.5.0
pydantic-settings==2.1.0

# CORS support
python-multipart==0.0.6

# System monitoring
psutil==5.9.6

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
aiosqlite==0.19.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Linting and formatting
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Environment variables
python-dotenv==1.0.0
